Remove triggered/dependent task distinction
This simplifies logic in several places where normal tasks are cast to
dependent (future) tasks, and the allocation logic as well. We might
want to add dependent tasks back later to save space, but this is
working well for now, and the overhead is not significant. This is still
a step in the right direction because we got rid of the triggered task
struct (and the associated bookkeeping).

Also removed unqualified [gs]et_current_finish functions, replacing with
direct field accesses instead. These functions weren't static, but also
weren't prefixed with hclib_, so it seemed easier to just drop them and
do the field accesses directly instead.